requirements.txt
pytest
allure-python-commons
Appium-Python-Client
selenium
openpyxl
allure-pytest